requirements.txt
amqp==2.3.2
aniso8601==3.0.2
astroid==1.6.5
billiard==3.5.0.4
celery==4.2.1
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.9.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
redis==2.10.6
six==1.11.0
vine==1.1.4
Werkzeug==0.14.1
wrapt==1.10.11
